Move reading of attributes files to config
