+++ commands/command_track.go
@@ -28,10 +28,6 @@ var (
 	trackNotLockableFlag    bool
 )
 
-const (
-	lockableAttrib = "lockable"
-)
-
 func trackCommand(cmd *cobra.Command, args []string) {
 	requireGitVersion()
 
@@ -46,7 +42,7 @@ func trackCommand(cmd *cobra.Command, args []string) {
 	}
 
 	lfs.InstallHooks(false)
-	knownPaths := findPaths()
+	knownPaths := config.GetAttributePaths()
 
 	if len(args) == 0 {
 		Print("Listing tracked paths")
@@ -87,7 +83,7 @@ ArgsLoop:
 		encodedArg := strings.Replace(pattern, " ", "[[:space:]]", -1)
 		lockableArg := ""
 		if trackLockableFlag { // no need to test trackNotLockableFlag, if we got here we're disabling
-			lockableArg = " " + lockableAttrib
+			lockableArg = " " + config.LockableAttrib
 		}
 
 		changedAttribLines[pattern] = fmt.Sprintf("%s filter=lfs diff=lfs merge=lfs -text%v\n", encodedArg, lockableArg)
@@ -188,69 +184,6 @@ ArgsLoop:
 			}
 		}
 	}
-}
-
-type mediaPath struct {
-	Path     string
-	Source   string
-	Lockable bool
-}
-
-func findPaths() []mediaPath {
-	paths := make([]mediaPath, 0)
-
-	for _, path := range findAttributeFiles() {
-		attributes, err := os.Open(path)
-		if err != nil {
-			continue
-		}
-
-		scanner := bufio.NewScanner(attributes)
-
-		for scanner.Scan() {
-			line := scanner.Text()
-			if strings.Contains(line, "filter=lfs") {
-				fields := strings.Fields(line)
-				relfile, _ := filepath.Rel(config.LocalWorkingDir, path)
-				pattern := fields[0]
-				if reldir := filepath.Dir(relfile); len(reldir) > 0 {
-					pattern = filepath.Join(reldir, pattern)
-				}
-				lockable := strings.Contains(line, lockableAttrib)
-				paths = append(paths, mediaPath{Path: pattern, Source: relfile, Lockable: lockable})
-			}
-		}
-	}
-
-	return paths
-}
-
-func findAttributeFiles() []string {
-	paths := make([]string, 0)
-
-	repoAttributes := filepath.Join(config.LocalGitDir, "info", "attributes")
-	if info, err := os.Stat(repoAttributes); err == nil && !info.IsDir() {
-		paths = append(paths, repoAttributes)
-	}
-
-	filepath.Walk(config.LocalWorkingDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if !info.IsDir() && (filepath.Base(path) == ".gitattributes") {
-			paths = append(paths, path)
-		}
-		return nil
-	})
-
-	// reverse the order of the files so more specific entries are found first
-	// when iterating from the front (respects precedence)
-	for i, j := 0, len(paths)-1; i < j; i, j = i+1, j-1 {
-		paths[i], paths[j] = paths[j], paths[i]
-	}
-
-	return paths
 }
 
 // blocklistItem returns the name of the blocklist item preventing the given
