--- locking/readonly.go
@@ -1,1 +1,0 @@
-package locking
